make a C version of SA1inittable and remove the asm version
No good reason for the asm version.  All it's doing is filling in some
arrays with opcodes.  SA1init can be written in C just like inittable
and inittablec.

This change also gets one step closer to being able to run this as a
native x86-64 binary.  The assembly SA1inittable function does some
stuff with the ES register which generates Illegal Instruction when
comipled as a 64-bit binary.